AiInternTask/requirements.txt
langchain
streamlit
chromadb
python-dotenv
PyPDF2
PyMuPDF
pytesseract
pillow
pdfplumber
openai
tqdm
sentence-transformers
unstructured
docx
langchain_community


langchain-chroma

pandas
requests

langchain-huggingface



langchain-core
torch
transformers

langchain_openai
python-docx
